Add docker container
Adds changes to run the dashboard using gunicorn within a
docker container to support more users. Adds a github
action to build the container and push to GHCR when
a release is created. Once a container is released we can
update the README to allow users to skip the build step.

See #40